arm64: add 64-bit time for all arm64 platform
Summary
  For ARM64 architecture, the arch timer is 64-bit,
the CONFIG_SYSTEM_TIME64 need to be enabled just like
x86_64 and risc-v 64

+++ arch/arm64/src/common/arm64_arch_timer.c
@@ -134,7 +134,7 @@ static int arm64_arch_timer_compare_isr(int irq, void *regs, void *arg)
 static int arm64_arch_timer_compare_isr(int irq, void *regs, void *arg)
 {
   uint64_t      curr_cycle;
-  uint32_t      delta_ticks;
+  uint64_t      delta_ticks;
   uint64_t      next_cycle;
 
   UNUSED(irq);
@@ -142,7 +142,7 @@ static int arm64_arch_timer_compare_isr(int irq, void *regs, void *arg)
   UNUSED(arg);
 
   curr_cycle    = arm64_arch_timer_count();
-  delta_ticks   = (uint32_t)((curr_cycle - last_cycle) / cycle_per_tick);
+  delta_ticks   = (curr_cycle - last_cycle) / cycle_per_tick;
 
   last_cycle += delta_ticks * cycle_per_tick;
